feat: recreate slash commands on server startup
